Moved pr 66 commits to project root in springboot-integration branch.
