--- com.dreamteam.database/buyer_event.csv
@@ -1,6 +1,0 @@
-date,email,shipping_address,product_id,quantity
-9/16/2020,buyingallyourthings@gmail.com,123 Coriander Way,2CE1RNHR6OV1,10
-9/16/2020,buyingallyourthings@gmail.com,123 Coriander Way,ZJ2VDLAXA5Y8,15
-9/16/2020,buyingallyourthings@gmail.com,123 Coriander Way,0S17TMY187O6,20
-9/16/2020,buyingallyourthings@gmail.com,123 Coriander Way,7DW0K069S8S6,25
-9/16/2020,buyingallyourthings@gmail.com,123 Coriander Way,GK9R7O4R0OXN,30
