--- com.dreamteam.database/customerHistory.csv
@@ -1,7 +1,0 @@
-Customer order,Date,Time
-00Smith,9-12-2020,1:00pm
-00Jones,9-13-2020,1:00pm
-00Alex,9-14-2020,1:00pm
-00Brian,9-15-2020,11:00am
-00Jacob,9-16-2020,5:00pm
-00John,9-17-2020,3:00pm
