--- com.dreamteam.database/customer_history.csv
@@ -1,2 +1,0 @@
-000Jon, 2020-09-22, 2020-09-22T11:57:45.383177100
-000BUY, 2020-09-22, 2020-09-22T11:58:06.367511100
