--- com.dreamteam.database/src/test/java/com/dreamteam/database/BuyerEventTest.java
@@ -1,11 +1,0 @@
-package com.dreamteam.database;
-
-public class BuyerEventTest {
-
-    // @Test
-    // public void testUpdateQuantity(){
-    //     //test method by comparing quantity values of products
-    //     //will need help with creating this test
-    // }
-
-}
