--- com.dreamteam.database/supplier_event.csv
@@ -1,6 +1,0 @@
-date,email,shipping_address,product_id,quantity
-09-15-20,000ALJ,2CE1RNHR6OV1,66
-09-15-20,000ALX,ZJ2VDLAXA5Y8,90
-09-15-20,000KVN,0S17TMY187O6,50
-09-15-20,000JON,7DW0K069S8S6,23
-09-15-20,000DOM,GK9R7O4R0OXN,70
